moved code into components
