+++ App.js
@@ -2,33 +2,27 @@ import {useState} from 'react';
 import { StyleSheet, Text, View, Button, TextInput, ScrollView, FlatList} from 'react-native';
 import { Colors } from 'react-native/Libraries/NewAppScreen';
 
+import GoalItem from './components/GoalItem'
+import GoalInput from './components/GoalInput'
+
 export default function App() {
-  const [enteredGoalText, setEnteredGoalText] = useState('');
+  
   const [courseGoals, setCourseGoals] = useState([]);
-
-  function goalInputerHandler(enteredText) {
-      setEnteredGoalText(enteredText);
-  }
+ 
 
 
-  function addGoalHandler() {
+
+  function addGoalHandler(enteredGoalText) {
     setCourseGoals((currentCourseGoals) => [...currentCourseGoals, {text: enteredGoalText, id: Math.random().toString()},
     ]);
   }
 
   return (
     <View style={styles.appContrainer}>
-     <View style={styles.inputContainer}>
-        <TextInput style={styles.textInput} placeholder='Your course goal!' onChangeText={goalInputerHandler} />
-        <Button title="Add Goal" onPress={addGoalHandler}/>
-     </View>
+      <GoalInput onAddGoal={addGoalHandler} />
      <View style={styles.goalsContainer} >
      <FlatList data={courseGoals} renderItem={(itemData)=> {
-      return (
-          <View style= {styles.goalItem}> 
-          <Text style={styles.goalText}> {itemData.item.text}</Text> 
-          </View>
-      );
+      return <GoalItem text={itemData.item.text}/>;
      }}
      keyExtract={(item, index) => {
       return item.id; }}
@@ -44,34 +38,8 @@ const styles = StyleSheet.create({
     paddingTop: 50,
     paddingHorizontal: 16
   },
-  inputContainer: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 24,
-    borderBottomWidth: 1,
-    borderBottomColor: '#cccccc'
-  },
-  textInput: {
-    borderWidth: 1,
-    borderColor:'#cccccc',
-    width: '70%',
-    marginRight: 8,
-    padding: 8
-  },
   goalsContainer: {
     flex: 5
   },
-  goalItem: {
-    margin: 8,
-    padding: 8,
-    borderRadius: 6,
-    backgroundColor: '#5e0acc',
-  },
-  goalText: {
-    color: 'white',
-  }
 
-  
-});
+ });
